# Improve shared informer notification dispatching