@ncdc 2017-07-15
Sorry for letting this slip through. I'll do another pass Monday.

On Sat, Jul 15, 2017 at 3:15 AM Mikhail Mazurskiy <notifications@github.com>
wrote:

> ptal @deads2k <https://github.com/deads2k> @ncdc <https://github.com/ncdc>
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/47045#issuecomment-315515969>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAABYplA1WAnxbQtMAoSyFTIS-a9dRDoks5sOGcVgaJpZM4NxVX0>
> .
>