@ncdc 2017-07-18
LGTM. I'd like another reviewer to take a look too.

@kubernetes/sig-api-machinery-pr-reviews @wojtek-t