@smarterclayton 2017-07-19
Are there performance numbers for this change?  I'd like to make sure this doesn't regress.