@ash2k 2017-07-19
@smarterclayton I'll add a benchmark for the listener.