@ash2k 2017-07-19
First number is `concurrencyLevel`:
```
Original:
1: 1000000	      1324 ns/op	     127 B/op	       2 allocs/op
5: 1000000	      1428 ns/op	     134 B/op	       2 allocs/op

PR:
1: 500000	      3139 ns/op	     101 B/op	       1 allocs/op
5: 500000	      3680 ns/op	     105 B/op	       1 allocs/op
```
So it is slower but does 1 less allocation. I don't know how important this slow down is - it probably is not. Also, this is mostly testing Go's concurrency primitives - in practice each listener will be doing something useful which will probably take much more time and amortize the cost.

p.s. I don't understand why my version is slower. Maybe the benchmark is wrong?