@deads2k 2017-07-19
> p.s. I don't understand why my version is slower. Maybe the benchmark is wrong?

Logically channels are going to require more lock acquisition.  





> So it is slower but does 1 less allocation. I don't know how important this slow down is

We push a lot of data through this path.  Every mutation to every resource in the entire system, multiple times in some cases.  @smarterclayton @wojtek-t