@wojtek-t 2017-07-19
Yeah - I also wanted to ask for benchmark but Clayton was first. That's unfortunately what I was expecting, to be honest.

> So it is slower but does 1 less allocation. I don't know how important this slow down is

I think this slow-down is pretty important. It's not 10% slow down - it's almost 3x slow down. As @deads2k already wrote this code path is very busy in pretty much all our components, so I think I'm against this change (though I think we should merge the benchmark as it may be useful in the future too).