@ash2k 2017-07-19
It will take 3 seconds to push 1 million messages vs 1.3 seconds. How big is this increase relatively to other work that needs to be done to process 1 million messages?

Anyway, if cpu consumption is more important in this case than readability of code I don't mind if this pr will be closed. I didn't expect it to be slower when I opened it so it's ok :)

p.s. I've tried to replace the `pendingNotifications` slice with a ring buffer slice but it is much worse for some reason. Haven't investigated yet. It's here if anyone is curious https://github.com/atlassian/kubernetes/commit/fdb9fd8cd6ff19d18ab163958ed2bfd032c528dc