@ash2k 2017-07-20
I profiled the benchmark and it is indeed the `select` statement that is causing the slowness. So I removed it from the `add()` method and also simplified the whole thing a bit.

Now it is a proper pipeline: informer starts the processor which starts the listeners, first the `pop()` method and then the `run()` method. `run()` method delivers notifications.

Shutdown is done in reverse order - controller sets the `stopped` flag to prevent addition of new listeners, then stops the processor. Processor closes the `addCh` which causes `pop()` to stop and close the `nextCh` which causes `run()` to stop.

```
1000000	      1451 ns/op	      96 B/op	       1 allocs/op
```

p.s. the additional allocation that is removed here was caused by the `blockingGet` anonymous function that is allocated in the `for` loop in `pop()`.

@deads2k @ncdc