@wojtek-t 2017-07-20
Those results seem much better - with such a small slow down and 1 allocation less, I think we can merge it. @smarterclayton - thoughts?